+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Facades/Meta/MetaNamespacerFacade.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GatoGraphQL\GatoGraphQL\Facades\Meta;
+
+use PoP\Root\App;
+use GatoGraphQL\GatoGraphQL\Meta\MetaNamespacerInterface;
+
+class MetaNamespacerFacade
+{
+    public static function getInstance(): MetaNamespacerInterface
+    {
+        /**
+         * @var MetaNamespacerInterface
+         */
+        $service = App::getContainer()->get(MetaNamespacerInterface::class);
+        return $service;
+    }
+}